rename file and add exports to style
