+++ src/components/robot-item.tsx
@@ -11,7 +11,7 @@ import { RobotInformation } from './robot-item-information';
 import * as RomiCore from '@osrf/romi-js-core-interfaces';
 import React from 'react';
 
-import OmniPanelStatusLabels from './omniPanelStatusLabels';
+import OmniPanelStatusLabels from './omni-panel-status-labels';
 
 export interface RobotItemProps extends Omit<ExpansionPanelProps, 'children'> {
   fleetName: string;
